Fixed WCTP021
The issue was due to the AI deciding that it wanted to start using curly quotes. I have made 2 changes, first is updating the system message 2 disallow the use of curly quotes. Then if the AI decides to still do it, I have adjusted runGeneration to replace curly quotes with smart quotes.