Completed CSVP049
Things are now setup so that everything that all the AI Environment Variables are retrieved through the AIConfig file. These are then injected into the services and therefore can be used everywhere. There are 2 exceptions, that cannot be avoided due to them not being possible to setup before the services begin.